Fix errors due to upgrades.
Either the new compiler or perhaps a TestNG upgrades made the TestNG
assertEquals(double, double, double) much more strict in that if one of
the arguments is not a double, there is no automatic promotion, so an
error is thrown.

The changes in the pom were mostly cosmetic.

+++ src/test/java/org/apache/datasketches/kll/DeprecatedAndMiscTest.java
@@ -59,9 +59,9 @@ public class DeprecatedAndMiscTest {
       kll.update(i);
     }
     final double eps = kll.getNormalizedRankError(false);
-    final double est = kll.getQuantile(0.5);
-    final double ub = kll.getQuantileUpperBound(0.5);
-    final double lb = kll.getQuantileLowerBound(0.5);
+    final float est = kll.getQuantile(0.5);
+    final float ub = kll.getQuantileUpperBound(0.5);
+    final float lb = kll.getQuantileLowerBound(0.5);
     assertEquals(ub, kll.getQuantile(.5 + eps));
     assertEquals(lb, kll.getQuantile(0.5 - eps));
     println("Ext     : " + est);
